backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
drf-yasg==1.21.7
django-allauth==0.57.0
PyJWT==2.8.0
pytest==7.4.4
pytest-django==4.7.0
coverage==7.4.1
requests==2.31.0
python-jose==3.3.0 